Moved ddtrace.compat to ddtrace.internal.compat, kept `public` but deprecated version.
